+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-# 11.1.1 - TBD
+# TBD
 ### Changed
 * Upgrade housekeeping to version 1.0.2
+
+# 11.1.1 - 2018-02-15
+### Fixed
+* Google FileSystem classes not being placed onto the mapreduce.application.classpath in S3MapReduceCp and DistCp mapreduce jobs
+### Changed
+* Google FileSystem and S3 FileSystems added to mapreduce.application.classpath in circus-train-gcp and circus-train-aws respectively
 
 # 11.1.0 - 2018-02-05
 ### Fixed